+++ projects/pip-webui2-layouts/src/lib/media/media.module.ts
@@ -3,19 +3,22 @@ import { BREAKPOINT } from '@angular/flex-layout';
 
 import { PipMediaService } from './shared/media.service';
 import { PIP_BREAKPOINTS } from './shared/breakpoints';
+import { ModuleWithProviders } from '@angular/compiler/src/core';
 
 const PipBreakPointsProvider = {
   provide: BREAKPOINT,
   useValue: [...PIP_BREAKPOINTS]
 };
 
-@NgModule({
-  declarations: [],
-  imports: [],
-  exports: [],
-  providers: [
-    PipMediaService,
-    PipBreakPointsProvider
-  ]
-})
-export class PipMediaModule { }
+@NgModule()
+export class PipMediaModule {
+  static forRoot(): ModuleWithProviders {
+    return {
+      ngModule: PipMediaModule,
+      providers: [
+        PipMediaService,
+        PipBreakPointsProvider
+      ]
+    };
+  }
+}