+++ projects/pip-webui2-layouts/src/lib/rightnav/rightnav.module.ts
@@ -1,13 +1,22 @@
-import { NgModule } from '@angular/core';
+import { NgModule, ModuleWithProviders } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { MatSidenavModule } from '@angular/material';
 
 import { PipRightnavService } from './shared/rightnav.service';
 
 @NgModule({
-  declarations: [],
-  imports: [],
-  exports: [],
-  providers: [
-    PipRightnavService
+  imports: [
+    CommonModule,
+    MatSidenavModule
   ]
 })
-export class PipRightnavModule { }
+export class PipRightnavModule {
+  static forRoot(): ModuleWithProviders {
+    return {
+      ngModule: PipRightnavModule,
+      providers: [
+        PipRightnavService
+      ]
+    };
+  }
+}